+++ threddsclient/__init__.py
@@ -1,2 +1,3 @@
+from .parser import *
 from .nodes import *
 from .catalog import *
